requirements.txt
langchain
langchain-community
langchain-anthropic
langchain-huggingface
langchain-text-splitters
lancedb
tantivy
sentence-transformers
python-dotenv
tqdm
fastapi
uvicorn
pandas
pyarrow
torch
torch-directml
numpy
jq
